refactor: add a base class LLMStreamService (#561)
* refactor: add a base class LLMStreamService

* refactor: change GeminiService to inherit from LLMStreamService

* perf: remove #available macos-12 in Gemini

* chore: update swift lint

* perf: remove unused swiftlint:disable

* perf: improve structure between LLMStreamService and BaseOpenAIService

* perf: make subclass must override properties availableModels, apiKey and endpoint

* perf: mark model as must be overridden
