+++ Easydict/Swift/Service/Caiyun/CaiyunResponse.swift
@@ -1,4 +1,3 @@
-// swiftlint:disable all
 //
 //  CaiyunResponse.swift
 //  Easydict
@@ -14,5 +13,3 @@ struct CaiyunResponse: Codable {
     var rc: Int
     var target: [String]
 }
-
-// swiftlint:enable all
