+++ Easydict/Swift/Service/Gemini/GeminiService.swift
@@ -6,15 +6,13 @@
 //  Copyright © 2024 izual. All rights reserved.
 //
 
-// swiftlint:disable all
-
 import Defaults
 import Foundation
 import GoogleGenerativeAI
 
 // TODO: add a LLM stream service base class, make both OpenAI and Gemini inherit from it.
 @objc(EZGeminiService)
-public final class GeminiService: QueryService {
+public final class GeminiService: LLMStreamService {
     // MARK: Public
 
     override public func serviceType() -> ServiceType {
@@ -29,21 +27,8 @@ public final class GeminiService: QueryService {
         NSLocalizedString("gemini_translate", comment: "The name of Gemini Translate")
     }
 
-    override public func supportLanguagesDictionary() -> MMOrderedDictionary<AnyObject, AnyObject> {
-        // TODO: Replace MMOrderedDictionary.
-        let orderedDict = MMOrderedDictionary<AnyObject, AnyObject>()
-        for language in EZLanguageManager.shared().allLanguages {
-            let value = language.rawValue
-            if !GeminiService.unsupportedLanguages.contains(language) {
-                orderedDict.setObject(value as NSString, forKey: language.rawValue as NSString)
-            }
-        }
-
-        return orderedDict
-    }
-
-    public override func isStream() -> Bool {
-        true
+    override public func queryTextType() -> EZQueryTextType {
+        [.translation]
     }
 
     override public func translate(
@@ -55,54 +40,45 @@ public final class GeminiService: QueryService {
         Task {
             do {
                 let translationPrompt = translationPrompt(text: text, from: from, to: to)
-                let prompt = QueryService.translationSystemPrompt +
+                let prompt = LLMStreamService.translationSystemPrompt +
                     "\n" + translationPrompt
-//                logInfo("gemini prompt: \(prompt)")
                 let model = GenerativeModel(
                     name: "gemini-pro",
                     apiKey: apiKey,
                     safetySettings: [
-                        GeminiService.harassmentSafety,
-                        GeminiService.hateSpeechSafety,
-                        GeminiService.sexuallyExplicitSafety,
-                        GeminiService.dangerousContentSafety,
+                        harassmentBlockNone,
+                        hateSpeechBlockNone,
+                        sexuallyExplicitBlockNone,
+                        dangerousContentBlockNone,
                     ]
                 )
 
+                result.isStreamFinished = false
+
+                var resultString = ""
+
                 // Gemini Docs: https://github.com/google/generative-ai-swift
-                if #available(macOS 12.0, *) {
-                    result.isStreamFinished = false
 
-                    var resultString = ""
-                    let outputContentStream = model.generateContentStream(prompt)
+                let outputContentStream = model.generateContentStream(prompt)
+                for try await outputContent in outputContentStream {
+                    guard let line = outputContent.text else {
+                        return
+                    }
+                    if !result.isStreamFinished {
+                        resultString += line
 
-                    for try await outputContent in outputContentStream {
-                        guard let line = outputContent.text else {
-                            return
-                        }
-                        if !result.isStreamFinished {
-                            resultString += line
-                            result.translatedResults = [resultString]
-                            await MainActor.run {
-                                throttler.throttle { [unowned self] in
-                                    completion(result, nil)
-                                }
+                        result.translatedResults = [resultString]
+                        await MainActor.run {
+                            throttler.throttle { [unowned self] in
+                                completion(result, nil)
                             }
                         }
                     }
-                    result.isStreamFinished = true
-                    completion(result, nil)
-                } else {
-                    // Gemini does not support stream in macOS 12.0-
-                    let outputContent = try await model.generateContent(prompt)
-                    guard let resultString = outputContent.text else {
-                        return
-                    }
-                    result.translatedResults = [resultString]
-                    await MainActor.run {
-                        completion(result, nil)
-                    }
                 }
+
+                result.isStreamFinished = true
+                result.translatedResults = [getFinalResultText(text: resultString)]
+                completion(result, nil)
             } catch {
                 /**
                  https://github.com/google/generative-ai-swift/issues/89
@@ -126,34 +102,32 @@ public final class GeminiService: QueryService {
 
     // MARK: Internal
 
-    let throttler = Throttler()
+    // https://ai.google.dev/available_regions
+    override var unsupportedLanguages: [Language] {
+        [
+            .persian,
+            .filipino,
+            .khmer,
+            .lao,
+            .malay,
+            .mongolian,
+            .burmese,
+            .telugu,
+            .tamil,
+            .urdu,
+        ]
+    }
+
+    // easydict://writeKeyValue?EZGeminiAPIKey=xxx
+    override var apiKey: String {
+        Defaults[.geminiAPIKey] ?? ""
+    }
 
     // MARK: Private
 
-    // https://ai.google.dev/available_regions
-    private static let unsupportedLanguages: [Language] = [
-        .persian,
-        .filipino,
-        .khmer,
-        .lao,
-        .malay,
-        .mongolian,
-        .burmese,
-        .telugu,
-        .tamil,
-        .urdu,
-    ]
-
     // Set Gemini safety level to BLOCK_NONE
-    private static let harassmentSafety = SafetySetting(harmCategory: .harassment, threshold: .blockNone)
-    private static let hateSpeechSafety = SafetySetting(harmCategory: .hateSpeech, threshold: .blockNone)
-    private static let sexuallyExplicitSafety = SafetySetting(harmCategory: .sexuallyExplicit, threshold: .blockNone)
-    private static let dangerousContentSafety = SafetySetting(harmCategory: .dangerousContent, threshold: .blockNone)
-
-    // easydict://writeKeyValue?EZGeminiAPIKey=xxx
-    private var apiKey: String {
-        Defaults[.geminiAPIKey] ?? ""
-    }
+    private let harassmentBlockNone = SafetySetting(harmCategory: .harassment, threshold: .blockNone)
+    private let hateSpeechBlockNone = SafetySetting(harmCategory: .hateSpeech, threshold: .blockNone)
+    private let sexuallyExplicitBlockNone = SafetySetting(harmCategory: .sexuallyExplicit, threshold: .blockNone)
+    private let dangerousContentBlockNone = SafetySetting(harmCategory: .dangerousContent, threshold: .blockNone)
 }
-
-// swiftlint:enable all
