+++ Easydict/Swift/Utility/Extensions/String/String+Extension.swift
@@ -18,7 +18,7 @@ extension String {
     func trim() -> String {
         trimmingCharacters(in: .whitespacesAndNewlines)
     }
-    
+
     /// Replace all newlines with whitespaces.
     /// For line breaks, currently macOS is `\n`, previously used `\r`, Windows is `\r\n`.
     func replacingNewlinesWithWhitespace() -> String {
